chore: fix user domain
